All the queries were extracted into an external file
