+++ users/src/modules/gamesave/data-access/gamesaveRepository.js
@@ -1,11 +1,7 @@
 const db = require('../../../db/db.js');
+const queries = require('./gameSavesQueries.js');
 
 const createGameSave = async (gameSaveData) => {
-    const query = `
-        INSERT INTO game_saves (match_id, move_number, player_id, move_coordinates, resulting_board_state)
-        VALUES ($1, $2, $3, $4, $5)
-        RETURNING *;
-    `;
     const values = [
         gameSaveData.matchId,
         gameSaveData.moveNumber,
@@ -14,37 +10,22 @@ const createGameSave = async (gameSaveData) => {
         gameSaveData.resultingBoardState
     ];
 
-    const { rows } = await db.query(query, values);
+    const { rows } = await db.query(queries.createGameSave, values);
     return rows[0];
 };
 
 const getGameSavesByMatchId = async (matchId) => {
-    const query = `
-        SELECT * FROM game_saves 
-        WHERE match_id = $1 
-        ORDER BY move_number ASC
-    `;
-    const { rows } = await db.query(query, [matchId]);
+    const { rows } = await db.query(queries.findGameSavesByMatch, [matchId]);
     return rows;
 };
 
 const getGameSaveByMatchIdAndMoveNumber = async (matchId, moveNumber) => {
-    const query = `
-        SELECT * FROM game_saves 
-        WHERE match_id = $1 AND move_number = $2
-    `;
-    const { rows } = await db.query(query, [matchId, moveNumber]);
+    const { rows } = await db.query(queries.findGameSavesByMatchAndMove, [matchId, moveNumber]);
     return rows[0];
 };
 
 const getLatestGameSaveByMatchId = async (matchId) => {
-    const query = `
-        SELECT * FROM game_saves 
-        WHERE match_id = $1 
-        ORDER BY move_number DESC 
-        LIMIT 1
-    `;
-    const { rows } = await db.query(query, [matchId]);
+    const { rows } = await db.query(queries.findLastGameSaveByMatch, [matchId]);
     return rows[0];
 };
