--- users/src/modules/user/domain/userResponseDTO.js
@@ -1,10 +1,0 @@
-const toUserResponseDto = (user) => {
-  return {
-    username: user.username,
-    email: user.email,
-    photo: user.photo,
-    nickname: user.nickname
-  };
-};
-
-module.exports = { toUserResponseDto };
